SizeOf.guesso now returns an int instead of a double

+++ main/src/main/scala/de/pfke/squeeze/zlib/data/length/digital/BitLengthIncludes.scala
@@ -11,7 +11,7 @@ trait BitLengthIncludes {
   }
 
   implicit class toBitLength(
-    in: Double
+    in: Int
     ) {
     def bit = BitLength(in)
     def bits: BitLength = bit
